Software Engineering Group Project/requirements.txt
brotli==1.0.9
ConfigParser==5.2.0
cryptography==35.0.0
Cython==0.29.26
dl==0.1.0
docutils==0.18.1
Flask==2.0.1
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==0.15.1
HTMLParser==0.0.2
importlib_metadata==4.10.0
ipaddr==2.2.0
Jinja2==3.0.1
keyring==23.5.0
lockfile==0.12.2
lxml==4.7.1
mypy_extensions==0.4.3
numpy==1.21.4
ordereddict==1.1
protobuf==3.19.3
pycryptodome==3.12.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyotp==2.6.0
python-dotenv==0.19.2
sshtunnel==0.4.0
typing_extensions==4.0.1
Werkzeug==2.0.2
wincertstore==0.2.1
WTForms==3.0.1
xmlrpclib==1.0.1
zipp==3.7.0
pycryptodome
wheel==0.37.0
setuptools
PyMySQL
python-dotenv
sshtunnel
pycryptodome
app~=0.0.1
cryptography
Pillow~=9.0.0
crypto~=1.4.1
SQLAlchemy~=1.4.28